Add files via upload
Trying to figure out why only the first image has an active modal. I tried a few things to fix it, but nothing works. Was wondering why modal only works for the first image. And here-in lies the danger of just using pre-written code instead of learning how to do it yourself. (But I'm going to watch a video I found to see how they work better!)